requirements.txt
streamlit
psycopg2-binary
pandas
python-dateutil
boto3
